core/vm: test the generated dispatch makes no real stack-helper calls. consolidated tests into one file
